第二回レビュー反映
wcクラスを定義して、そこでユーザーからの入力に応じたword countの処理を呼び出すように変更
wcクラスで表示する出力文字列を組み立て、mainでputsするように変更
requireを全て一括でmain.rbで呼び出していたため、必要なクラスに書くように修正
