+++ 09.wc_object/lib/input_calculator.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative 'calc'
 class InputCalculator
   def initialize(option, input)
     @option = option
